print number

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -7,7 +7,7 @@
 
 void print_number(int n)
 {
-	unsigned int long x = 1, y = n, res;
+	unsigned int x = 1, y = n, res;
 
 	if (n == 0)
 		_putchar('0');
